// frontend/src/app/shared/generated/models/schedule-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ScheduleDto {
  analysisType?: (string | null);
  assignedToUserId?: (number | string | null);
  exclusionRules?: (string | null);
  id: (number | string);
  isActive: boolean;
  name: string;
  recurrencePattern?: (string | null);
  rowVersion: string;
  shiftPattern: string;
  site: string;
}